+++ src/ILedController.h
@@ -6,7 +6,7 @@
 #define LOG_TAG_LED "ILedController"
 
 // Pattern types supported:
-enum  Pattern { NONE, RAINBOW_CYCLE, THEATER_CHASE, COLOR_WIPE, SCANNER, FADE, RESCUE_FLASH_LIGHT };
+enum  Pattern { NONE, RAINBOW_CYCLE, THEATER_CHASE, COLOR_WIPE, CYLON, FADE, RESCUE_FLASH_LIGHT };
 // Patern directions supported:
 enum  Direction { FORWARD, REVERSE };